+++ src/modules/voxelformat/QBCLFormat.h
@@ -10,14 +10,13 @@ namespace voxel {
 
 /**
  * @brief Qubicle project file (qbcl) format.
- *
- * Not yet implemented
  */
 class QBCLFormat : public Format {
 private:
-	bool readMatrix(io::SeekableReadStream &stream);
-	bool readModel(io::SeekableReadStream &stream);
-	bool readCompound(io::SeekableReadStream &stream);
+	bool readMatrix(const core::String &filename, io::SeekableReadStream& stream, SceneGraph& sceneGraph, int parent, const core::String &name);
+	bool readModel(const core::String &filename, io::SeekableReadStream& stream, SceneGraph& sceneGraph, int parent, const core::String &name);
+	bool readCompound(const core::String &filename, io::SeekableReadStream& stream, SceneGraph& sceneGraph, int parent, const core::String &name);
+	bool readNodes(const core::String &filename, io::SeekableReadStream& stream, SceneGraph& sceneGraph, int parent);
 public:
 	image::ImagePtr loadScreenshot(const core::String &filename, io::SeekableReadStream& stream) override;
 	bool loadGroups(const core::String &filename, io::SeekableReadStream& stream, SceneGraph& sceneGraph) override;