+++ src/modules/voxelformat/tests/QBCLFormatTest.cpp
@@ -11,7 +11,7 @@ namespace voxel {
 class QBCLFormatTest: public AbstractVoxFormatTest {
 };
 
-TEST_F(QBCLFormatTest, DISABLED_testLoad) {
+TEST_F(QBCLFormatTest, testLoad) {
 	QBCLFormat f;
 	std::unique_ptr<RawVolume> volume(load("qubicle.qbcl", f));
 	ASSERT_NE(nullptr, volume) << "Could not load qbcl file";